mmngr_drv: mmngr: For an address in CMA for Lossy, do address conversion
When IPMMU is enable, it is necessary to do address conversion from
virtual address to physical address.

A physical address is calculated based on a virtual address and
CMA_LOSSY_VIRT_BASE_ADDR which is the start address to store
Lossy compressed data.

An address for Common CMA, don't need to do address conversion
because physical address is equal to virtual address.
However, an address for CMA for MMP and CMA for Lossy,
we need to do address conversion.

Signed-off-by: Tomoharu Fukawa <tomoharu.fukawa.eb@renesas.com>