Revert "mmngr_drv: mmngr: Add UTLB for FCP-VI (VSPI) in r8a7795 WS2.0"
This reverts 'commit 027a6052ace5 ("mmngr_drv: mmngr: Add UTLB for
FCP-VI (VSPI) in r8a7795 WS2.0")'.

Due to IPMMU restriction in r8a7795 ES2.0, allow only
one IP master to enable IPMMU(PMB) in the same IPMMU cache.

In case of IPMMU-VP0(VP1), it has the UTLBs of FCP-F, FCP-VB and FCP-VI.
We select the IPMMU support for FCP-F for OMX midlleware.
So this patch disables that for FCP-VI.

Signed-off-by: Tomoharu Fukawa <tomoharu.fukawa.eb@renesas.com>