mmngr_drv: mmngr: Change the translation policy in Lossy area on r8a7795 ES2.0
This patch is for workaround.

Due to IPMMU restriction in r8a7795 ES2.0, we only enable the IPMMU support
for FCP-F for OMX midlleware and disable that for FCP-VB(FCP-VI) in the
same IPMMU cache.

In order for VSPB(VSPI) to read Lossy data which is located in Legacy area
without the PMB support, a virtual address in Lossy area should be the same
as a physical of that.

Change the address translation policy of Lossy area:

<Before>
    H3 WS1.x : The VA of Lossy area is 0x20000000.
    H3 ES2.0 : The VA of that is 0x20000000.
    M3 WS1.0 : The VA of that is 0x20000000.

<After>
    H3 WS1.x : The VA of Lossy area is 0x20000000.
    H3 ES2.0 : The VA of that is same as PA.
    M3 WS1.0 : The VA of that is 0x20000000.

Signed-off-by: Tomoharu Fukawa <tomoharu.fukawa.eb@renesas.com>