mmngr_drv: mmngr: handle invalid translation
When IPMMU support is enabled and address transformation involves a
translation table, the input physical address may not match any section
of the transformation table. In such case, the default zero value of
the virtual address is returned and propagated to the user.
The consequences of this error might be visible only in the userspace
and finding the root cause may require some effort. It is proposed to
handle translation faults as earlier as possible and return an error to
a client.

Since all values in the 32-bit range (including 0x00000000 and
0xffffffff) are valid output values an additional variable should be
used to store a translation result. It is proposed to add an additional
(output) function parameter of the boolean type.

Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>