mmngr_drv: mmngr: set 32-bit DMA mask for M3N and E3
With the current implementation, the MMNGR sets a 40-bit DMA bitmask
(when IPMMU is enabled) regardless of the actual SoC. At the same
time, M3N and E3 don't have 40-bit address space. IPMMU MMU supports
physical address space in the legacy area.
It is proposed to make changes that allow using the correct DMA
bitmask for all devices.

Signed-off-by: Oleksii Khramkov <oleksii.khramkov@globallogic.com>
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>